+++ src/collective/documentgenerator/content/style_template.py
@@ -33,4 +33,3 @@ class StyleTemplate(Item):
     """
 
     implements(IStyleTemplate)
-    meta_type = 'StyleTemplate'