+++ docs/CHANGES.rst
@@ -9,6 +9,8 @@ Changelog
   [sgeulette]
 - Return generation context from rendering methods to use it in tests
   [sgeulette]
+- Added unit testing for do_rendering feature
+  [odelaere]
 
 1.0.5 (2017-03-10)
 ------------------