+++ docs/CHANGES.rst
@@ -11,6 +11,8 @@ Changelog
   [sgeulette]
 - Added unit testing for do_rendering feature
   [odelaere, sgeulette]
+- Improved validation for ConfigurablePodTemplate
+  [odelaere]
 
 1.0.5 (2017-03-10)
 ------------------